+++ viewer/src/logic/utils.ts
@@ -1,0 +1,11 @@
+import Engine, { Phase } from "@gaia-project/engine";
+
+export function phaseBeforeSetupBuilding(data: Engine): boolean {
+  return (
+    data.phase === Phase.SetupInit ||
+    data.phase === Phase.SetupBoard ||
+    data.phase === Phase.SetupFaction ||
+    data.phase === Phase.SetupAuction
+  );
+}
+
